requirements/requirements-repro.txt
absl-py==2.3.1
accelerate==1.9.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
async-timeout==5.0.1
attrs==25.3.0
blinker==1.9.0
blobfile==3.0.0
brax==0.12.4
certifi==2019.11.28
cfgv==3.4.0
chardet==3.0.4
charset-normalizer==3.4.2
chex==0.1.90
click==8.1.8
click-option-group==0.5.7
cloudpickle==3.1.1
colorama==0.4.6
colorful==0.5.7
colorlog==6.9.0
contourpy==1.3.2
cycler==0.12.1
datasets==4.0.0
dbus-python==1.2.16
Deprecated==1.2.18
dill==0.3.8
distlib==0.4.0
dm-tree==0.1.9
dotty-dict==1.3.1
etils==1.13.0
exceptiongroup==1.3.0
fairscale==0.4.13
Farama-Notifications==0.0.4
fastapi==0.116.1
filelock==3.13.1
fire==0.7.0
Flask==3.1.1
flask-cors==6.0.1
flatbuffers==25.2.10
flax==0.10.7
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.45
glfw==2.9.0
graphviz==0.21
grpcio==1.74.0
gym==0.26.2
gym-notices==0.1.0
Gymnasium==0.26.3
gymnasium-notices==0.0.1
hf-xet==1.1.5
hf_transfer==0.1.9
huggingface-hub==0.34.1
humanize==4.12.3
identify==2.6.12
idna==2.8
imageio==2.37.0
importlib_resources==6.5.2
iniconfig==2.1.0
iotop==0.6
islpy==2025.2.5
itsdangerous==2.2.0
#jax==0.6.2
#jax-cuda12-pjrt==0.6.2
#jax-cuda12-plugin==0.6.2
#jaxlib==0.6.2
jaxopt==0.8.5
Jinja2==3.1.4
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kagglehub==0.3.12
kiwisolver==1.4.8
lazy_loader==0.4
lxml==6.0.0
lz4==4.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.3
mdurl==0.1.2
ml_collections==1.1.0
ml_dtypes==0.5.1
mock==5.2.0
mpmath==1.3.0
msgpack==1.1.1
mujoco==3.3.4
mujoco-mjx==3.3.4
multidict==6.6.3
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.3
nodeenv==1.9.1
numpy==2.2.6
nvidia-cublas-cu12==12.8.3.14
nvidia-cuda-cupti-cu12==12.8.57
nvidia-cuda-nvcc-cu12==12.9.86
nvidia-cuda-nvrtc-cu12==12.8.61
nvidia-cuda-runtime-cu12==12.8.57
nvidia-cudnn-cu12==9.11.0.98
#nvidia-cudnn-cu12==9.7.1.26
nvidia-cufft-cu12==11.3.3.41
nvidia-cufile-cu12==1.13.0.11
nvidia-curand-cu12==10.3.9.55
nvidia-cusolver-cu12==11.7.2.55
nvidia-cusparse-cu12==12.5.7.53
nvidia-cusparselt-cu12==0.6.3
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.8.61
nvidia-nvshmem-cu12==3.3.9
nvidia-nvtx-cu12==12.8.55
omegaconf==2.3.0
opt_einsum==3.4.0
optax==0.2.5
optree==0.17.0
orbax-checkpoint==0.11.19
packaging==25.0
pandas==2.3.1
pickle5==0.0.11
pillow==11.0.0
platformdirs==4.3.8
pluggy==1.6.0
pre_commit==4.2.0
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
PuLP==3.2.1
pyarrow==21.0.0
pycryptodomex==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyGObject==3.36.0
PyOpenGL==3.1.9
pyparsing==3.2.3
pytest==8.4.1
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.9.0.post0
python-gitlab==6.2.0
python-semantic-release==10.2.0
pytz==2025.2
ray==2.37.0
redis==6.2.0
referencing==0.36.2
regex==2024.11.6
requests-toolbelt==1.0.0
requests-unixsocket==0.2.0
rich==14.1.0
rl_games==1.6.1
rpds-py==0.26.0
safetensors==0.5.3
sample-factory==2.1.1
scikit-image==0.25.2
scikit-optimize==0.10.2
scipy==1.15.3
sentencepiece==0.2.0
sentry-sdk==2.33.2
shellingham==1.5.4
signal-slot-mp==1.0.5
simplejson==3.20.1

six==1.17.0
smmap==5.0.2
starlette==0.47.2
sympy==1.13.3
tensorboard==2.20.0
tensorboardX==2.6.4
tensorstore==0.1.76
termcolor==3.1.0
tifffile==2025.5.10
tiktoken==0.9.0
tokenizers==0.21.2
tomli==2.2.1
tomlkit==0.13.3
toolz==1.0.0
#torch==2.7.1+cu128
torchao==0.12.0
torchdata==0.11.0
torchtune==0.6.1
#torchvision==0.22.1+cu128
tqdm==4.67.1
treescope==0.1.9
trimesh==4.7.1
triton==3.3.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
uvicorn==0.35.0
virtualenv==20.32.0
wandb==0.21.0
Werkzeug==3.1.3
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
#urllib3==2.5.0
#requests==2.32.4
#PyYAML==6.0.2
